requirements.txt
python-telegram-bot>=22.3
aiosqlite
